chore: move `RenderedModule` to the `types` folder
