+++ crates/rolldown_binding/src/lib.rs
@@ -5,5 +5,6 @@ static ALLOC: mimalloc::MiMalloc = mimalloc::MiMalloc;
 pub mod bundler;
 pub mod options;
 pub mod output;
+pub mod types;
 pub mod utils;
 scoped_tls::scoped_thread_local!(static NAPI_ENV: napi::Env);
