+++ crates/rolldown_binding/src/types/mod.rs
@@ -1,0 +1,1 @@
+pub mod binding_rendered_module;
